Rename packeges to lower case and add GenericDAO interface
