+++ src/database/DBConnection.java
@@ -1,13 +1,13 @@
+package database;
+
 import java.sql.Connection;
 import java.sql.DriverManager;
 import java.sql.SQLException;
 
-public class DBConnection_example {
-    // Rename this file to DBConnection
-    // Insert your data for connection to database
-    private static final String URL = "jdbc:postgresql://localhost:your_port/your_db";
-    private static final String USER = "your_user";
-    private static final String PASSWORD = "user_password";
+public class DBConnection {
+    private static final String URL = "jdbc:postgresql://localhost:5432/docsgrid";
+    private static final String USER = "postgres";
+    private static final String PASSWORD = "1212";
 
     public static Connection connect() {
         try {
