+++ src/entities/Company.java
@@ -1,0 +1,4 @@
+package entities;
+
+public class Company {
+}
