+++ src/entities/enums/TypeOfUser.java
@@ -1,4 +1,4 @@
-package entities.Enums;
+package entities.enums;
 
 public enum TypeOfUser {
     EMPLOYEE,
